requirements.txt
#### Dependencies for PhiloLogic4 ####
lxml
ujson
python-levenshtein
